requirements.txt
streamlit
replicate
Pillow
requests
python-dotenv